tests/wpt/tests/css/css-view-transitions/navigation/with-types/at-rule-last-rule-wins.html

<!DOCTYPE html>
<title>View transitions: Algorithms resolve against last @view-transition rule</title>
<link rel="help" href="https://drafts.csswg.org/css-view-transitions-2/#resolve-view-transition-rule">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  @view-transition {
    navigation: auto;
  }
  @view-transition {
    types: foo;
  }
</style>
<script>
const params = new URLSearchParams(location.search);

switch (params.get("mode") || "test") {
case "test":
  promise_test(async t => {
    const event = await new Promise(resolve => {
      window.did_reveal = e => { resolve(e) };
      const popup = window.open("?mode=old");
      t.add_cleanup(() => popup.close());
    });

    assert_equals(event.viewTransition, null,
        "Transition must be skipped: last rule has no navigation:auto.");
  });
  break;
case "old":
  onload = () => requestAnimationFrame(() => requestAnimationFrame(() => {
      location.replace('?mode=new');
  }));
  break;
case "new":
  onpagereveal = e => window.opener.did_reveal(e);
  break;
}
</script>
